back/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
branca==0.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dependencies==7.1.7
Django==3.2.9
django-cors-headers==3.10.0
django-extensions==3.1.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
entrypoints==0.4
executing==0.8.3
folium==0.12.1.post1
fonttools==4.29.1
h11==0.13.0
idna==3.3
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
selenium==4.1.3
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.2
testpath==0.6.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wsproto==1.1.0
